requirements.txt
importlib-metadata==8.0.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.collections==5.1.0
loguru==0.7.2
openai==1.51.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pip-chill==1.0.3
platformdirs==4.2.2
python-dotenv==1.0.1
pyyaml==6.0.2
tomli==2.0.1
